restructured, applied black
